nerves_system_br 1.13.3; linux 5.4
This updates nerves_system_br to 1.13.3 to bring in updated Raspberry Pi
firmware and userland support. Consequently, this makes it possible to
update the Linux kernel from 4.19 to 5.4 to match Raspbian.

One consequence of the update is that the WiFi device drivers changed
between kernel versions and this switches where their firmware is pulled
from since the rpi-wifi-firmware firmware was out-of-date.