Support USB gadget mode on the USB-C port
This lets you connect a Pi4 to your computer and communicate with it
over the USB cable like the Nerves RPi0 and RPi3A ports. Your computer
must be able to supply enough power through USB for this to work since
the Pi4 needs quite a bit more than the others to be happy.

This removes arbitrary USB gadget configurations since the intended use
for this is to start the USB-C port up as a gadget Ethernet device. For
that to work, gadget Ethernet has to be started early and the fool-proof
way to do that is to configure the kernel to do it. If you want to make
USB-C look like a keyboard, MIDI, etc. device then the Linux
configuration needs to be changed. It's not a runtime option.

Fixes #18.